+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Buildbotics CNC Controller Firmware Changelog
  - Fixed zeroing with non-zero offset when unhomed. #211
  - Handle file paths uploaded from Windows correctly. #212
  - Don't retain estop state through reboot.
+ - Log when RPi gets hot.
+ - Support Modbus multi-write mode.
+ - Added support for Nowforever VFDs.
 
 ## v0.4.6
  - Fixed a rare ``Negative s-curve time`` error.